refactor local download code into utils
Now code can be used as a utility across files
